DATAHUB-1627/1652: Add enrolment import field remove_role
The remove_role field is valid for enrolment_update.
It can have preset values: '1', 'yes', 'true' - if the role
to remove is unambiguous - and 'all', or it maybe a valid
role shortname already assigned to the user on the context.
It is disabled when set to: '0', 'no' 'false' or left empty.